pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.4.77",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.1",
]
